// FE-admin/src/types/address.model.ts
export interface AddressModel {
  phoneCode: string | undefined;
  country: string | undefined;
  category: string | undefined;
  firstName: string | undefined;
  lastName: string | undefined;
  postalCode: string | undefined;
  address: string | undefined;
  phoneNumber: string | undefined;
  province: string | undefined;
  provinceEng: string | undefined;
  district: string | undefined;
  districtEng: string | undefined;
  subDistrict: string | undefined;
  subDistrictEng: string | undefined;
  default: boolean | undefined;
}